// src/api/schemas/bookmarkFolderDetailsSchema.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Keyo API
 * API for Keyo Polls app
 * OpenAPI spec version: 1.0.0
 */
import type { BookmarkFolderDetailsSchemaContentType } from './bookmarkFolderDetailsSchemaContentType';
import type { BookmarkFolderDetailsSchemaPrice } from './bookmarkFolderDetailsSchemaPrice';

/**
 * Detailed schema for bookmark folder with all information
 */
export interface BookmarkFolderDetailsSchema {
  id: number;
  folder_id: string;
  slug: string;
  name: string;
  description: string;
  color: string;
  access_level: string;
  access_level_display: string;
  price?: BookmarkFolderDetailsSchemaPrice;
  is_todo_folder: boolean;
  content_type?: BookmarkFolderDetailsSchemaContentType;
  is_private: boolean;
  is_public: boolean;
  is_paid: boolean;
  bookmark_count: number;
  created_at: string;
  updated_at: string;
}