// src/api/schemas/todoBookmarkStatusResponseSchema.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Keyo API
 * API for Keyo Polls app
 * OpenAPI spec version: 1.0.0
 */
import type { TodoBookmarkStatusResponseSchemaFolderId } from './todoBookmarkStatusResponseSchemaFolderId';
import type { TodoBookmarkStatusResponseSchemaFolderName } from './todoBookmarkStatusResponseSchemaFolderName';

/**
 * Response schema for todo bookmark status
 */
export interface TodoBookmarkStatusResponseSchema {
  todo_id: number;
  is_bookmarked: boolean;
  folder_id?: TodoBookmarkStatusResponseSchemaFolderId;
  folder_name?: TodoBookmarkStatusResponseSchemaFolderName;
}